+++ src/core/kv-storage/engines/string/const.ts
@@ -1,0 +1,14 @@
+/*!
+ * V4Fire Core
+ * https://github.com/V4Fire/Core
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Core/blob/master/LICENSE
+ */
+
+import type { DataSeparators } from 'core/kv-storage/engines/string/interface';
+
+export const defaultDataSeparators: DataSeparators = {
+	chunk: '{{#}}',
+	record: '{{.}}'
+};