+++ src/core/kv-storage/engines/string/index.ts
@@ -1,0 +1,15 @@
+/*!
+ * V4Fire Core
+ * https://github.com/V4Fire/Core
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Core/blob/master/LICENSE
+ */
+
+import StringEngine from 'core/kv-storage/engines/string/engine';
+
+export { StringEngine as default };
+
+export const
+	syncSessionStorage = new StringEngine(),
+	asyncSessionStorage = syncSessionStorage;